packages/sacred-parallelism/pyproject.toml
[project]
name = "sacred-parallelism"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "GBisi", email = "giuseppe.bisicchia@phd.unipi.it" }
]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "sacred>=0.8.7",
    "setuptools>=80.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
